% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{background_df}
\alias{background_df}
\title{Simulated Background Data for cPCA and scPCA}
\format{
A simple \code{data.frame}.
}
\usage{
data(background_df)
}
\description{
The background data consisting of 400 observations and 30 variables was
simulated as follows:
\itemize{
  \item Each of the first 10 variables was drawn from $N(0, 10)$
  \item Variables 11 through 20 were drawn from $N(0, 3)$
  \item Variables 21 through 30 were drawn from $N(0, 1)$
}
}
\examples{
data(background_df)
}
\keyword{datasets}
