% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCPCA.R
\name{bpFitCPCA}
\alias{bpFitCPCA}
\title{Contrastive Principal Component Analysis in Parallel}
\usage{
bpFitCPCA(target, center, scale, c_contrasts, contrasts, n_eigen,
  num_medoids)
}
\arguments{
\item{target}{The target data set.}

\item{center}{A \code{logical} indicating whether the data sets' columns
should be centered so as to have mean zero.}

\item{scale}{A \code{logical} indicating whether the data sets' columns
should be re-scaled to have unit variance.}

\item{c_contrasts}{List of of contrastive covariances.}

\item{contrasts}{Vector of contrastive parameter values used to compute the
contrastive covariances,}

\item{n_eigen}{The number of contrastive principal components to compute.}

\item{num_medoids}{The number of medoids to consider.}
}
\value{
A list of lists containing the cPCA results for each contrastive
  parameter deemed to be a medoid.
  \itemize{
    \item rotation - the list of matrices of variable loadings
    \item x - the list of rotated data, centred and scaled if requested,
    multiplied by the rotation matrix
    \item contrast - the list of contrastive parameters
    \item penalty - set to zero, since the loadings are not penalized in cPCA
  }
}
\description{
Given target and background dataframes or matrices, \code{cPCA}
  will perform contrastive principal component analysis (cPCA) of the target
  data for a given number of eigenvectors and a vector of real valued
  contrast parameters. This is identical to the implementation of cPCA
  method by Abid et al. \insertRef{abid2017contrastive}{scPCA}. Analogous
  to \code{\link{fitCPCA}}, but replaces all \code{lapply} calls by
  \code{\link[BiocParallel]{bplapply}}.
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
