% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArgs.R
\name{checkArgs}
\alias{checkArgs}
\title{Check Arguments passed to the scPCA Function}
\usage{
checkArgs(target, background, center, scale, n_eigen, contrasts, penalties)
}
\arguments{
\item{target}{The target (experimental) data set, in a standard format such
as a \code{data.frame} or \code{matrix}.}

\item{background}{The background data set, in a standard format such as a
\code{data.frame} or \code{matrix}.}

\item{center}{A \code{logical} indicating whether the target and background
data sets should be centered to mean zero.}

\item{scale}{A \code{logical} indicating whether the target and background
data sets should be scaled to unit variance.}

\item{n_eigen}{A \code{numeric} indicating the number of eigenvectors to be
computed.}

\item{contrasts}{A \code{numeric} vector of the contrastive parameters.}

\item{penalties}{A \code{numeric} vector of the penalty terms.}
}
\value{
Whether all argument conditions are satisfied
}
\description{
Checks whether or not the all arguments in the \code{scPCA}
  functions are input properly.
}
\keyword{internal}
