% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceMatrix.R
\name{coerceMatrix}
\alias{coerceMatrix}
\title{Coerce Matrix Objects to Base Matrix Objects}
\usage{
coerceMatrix(data)
}
\arguments{
\item{data}{The data to be coerced to a matrix object.}
}
\value{
returns a coerced matrix object.
}
\description{
Coerces an object from the Matrix class to a base matrix object.
}
