% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastiveCov.R
\name{contrastiveCov}
\alias{contrastiveCov}
\title{Contrastive Covariance Matrices}
\usage{
contrastiveCov(target, background, contrasts, center, scale)
}
\arguments{
\item{target}{The target data set}

\item{background}{The bacground data set}

\item{contrasts}{The vector of contrastive parameters}

\item{center}{Whether the data sets' columns should be centered}

\item{scale}{Whetehr the data sets' columns should be scaled to have variance
1}
}
\value{
A list of contrastive covariance matrices. Each element has an
  associated contrastive parameter in the \code{contrasts} vector.
}
\description{
Compute the list of contrastive covariance matrices.
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
