% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMat.R
\name{covMat}
\alias{covMat}
\title{Compute Sample Covariance Matrix}
\usage{
covMat(data, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{The data with which to compute the sample covariance matrix.}

\item{center}{A \code{logical} indicating whether the data sets' columns
should be centered so as to have mean zero.}

\item{scale}{A \code{logical} indicating whether the data sets' columns
should be re-scaled to have unit variance.}
}
\value{
the covariance matrix of the data.
}
\description{
\code{covMat} computes the sample covariance matrix of a data
  set. If a variable in the dataset has zero variance, then its corresponding
  row and column in the covariance matrix are zero vectors.
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
