% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPCA.R
\name{cPCA}
\alias{cPCA}
\title{Constrastive Principal Component Analysis}
\usage{
cPCA(target, background, center = TRUE, scale = TRUE, num_eigen = 2,
  contrasts = exp(seq(log(0.1), log(1000), length.out = 40)),
  num_medoids)
}
\arguments{
\item{target}{The target data. Either a numeric dataframe or a matrix with
observations as rows and features as columns.}

\item{background}{The background data. Either a numeric dataframe or a matrix
with observations as rows and features as columns. The number of features
must match the number of features in the target data.}

\item{center}{Whether the target and background data should have their
columns' centered. Defaults to \code{TRUE}.}

\item{scale}{Whether the target and background data should have their
columns' scaled. Defaults to \code{TRUE}.}

\item{num_eigen}{The number of contrastive principal components to compute.
Must be a non-negative integer between 1 and the number of columns in the
target data. Default is 2.}

\item{contrasts}{The numeric vector of the contrastive parameters. Each
element must be a unique non-negative real number. Defaults to 40
logarithmically spaced values between 0.1 and 1000.}

\item{num_medoids}{The number of medoids to select during the spectral
clustering of the constrastive subspaces. Defaults to
\code{round(length(contrasts)/5)} if there are at least 5 contrastive
parameters, otherwise \code{length(contrasts)-1}. If there is a single
contrastive parameter, then \code{num_medoids} should not be defined.}
}
\value{
A list containing the vector of contrast parameters that were
  selected as medoids by spectral clustering, the list of eigenvector
  matrices associated with each of these contrast parameters and the list
  of reduced-dimension target data associated with these contrast parameters.
}
\description{
Given target and background dataframes or matrices, \code{cPCA}
  will perform the contrastive principal component analysis of the target
  data for a given number of eigenvectors and a vector of real valued
  contrast parameters. For more information on this method, see
  \href{https://www.nature.com/articles/s41467-018-04608-8#ref-CR29}{Abid et al.}.
}
\examples{
cPCA(
  target = toy_df[, 2:31],
  background = background_df
)
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
