% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvSPC.R
\name{cvSPC}
\alias{cvSPC}
\title{Cross-Validation Scheme for SPC on Target Data}
\usage{
cvSPC(target, c_proj, cov_mat, penalties, n_eigen, n_iter, n_folds, v)
}
\arguments{
\item{target}{The target data set.}

\item{c_proj}{The optimal contrastive projection of the target data.}

\item{cov_mat}{The optimal contrastive covariance matrix.}

\item{penalties}{The vector of penalties for the SPC.}

\item{n_eigen}{The number of components to calculate.}

\item{n_iter}{Number of iterations to be performed.}

\item{n_folds}{The number of folds used for CV.}

\item{v}{The first right singular vector of \code{cov_mat}.}
}
\value{
The list of optimal penalty terms based on cross-validated mean
  squared error:
  \itemize{
    \item best - The optimal L1 penalty based of CV MSE
    \item best1se - The sparsest penalty within 1 standard error of CV MSE
  }
}
\description{
Cross-Validation Scheme for SPC on Target Data
}
\details{
This function is a modified version of the \code{\link[PMA]{SPC.cv}}
  function. The cross-validation procedure is described in
  \insertRef{WittenPMD2009}{scPCA}.
}
