% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContrast.R
\name{fitContrast}
\alias{fitContrast}
\title{Identify the Optimal Contrastive Parameter}
\usage{
fitContrast(target, center = TRUE, scale = TRUE, c_contrasts,
  contrasts, num_eigen, centers, iter_max = 100, ...)
}
\arguments{
\item{target}{The target data set.}

\item{center}{A \code{logical} indicating whether the data sets' columns
should be centered so as to have mean zero.}

\item{scale}{A \code{logical} indicating whether the data sets' columns
should be re-scaled to have unit variance.}

\item{c_contrasts}{List of of contrastive covariances.}

\item{contrasts}{Vector of contrastive parameter values used to compute the
contrastive covariances,}

\item{num_eigen}{The number of contrastive principal components to compute.}

\item{centers}{The number of centers to use in the clustering algorithm.}

\item{...}{Additional arguments to pass to the clustering algorithm and the
objective function.}
}
\value{
Returns the optimal covariance matrix, contrastive parameter,
  loadings and projection of the target data based on the results of
  evaluating a clustering algorithm based on average silhouette width metric.
}
\description{
This function is used to automatically select the optimal
  contrastive parameter based on k-means and average silhouette width.
}
