% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_contrastiveCov.R
\name{future_contrastiveCov}
\alias{future_contrastiveCov}
\title{Contrastive Covariance Matrices}
\usage{
future_contrastiveCov(target, background, contrasts, center, scale)
}
\arguments{
\item{target}{The target data set.}

\item{background}{The background data set.}

\item{contrasts}{The vector of contrastive parameters.}

\item{center}{A \code{logical} indicating whether the data sets' columns
should be centered so as to have mean zero.}

\item{scale}{A \code{logical} indicating whether the data sets' columns
should be re-scaled to have unit variance.}
}
\value{
A list of contrastive covariance matrices. Each element has an
  associated contrastive parameter in the \code{contrasts} vector.
}
\description{
Compute the list of contrastive covariance matrices.
Paralellized with the \code{future.apply} package.
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
