% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projGridCP.R
\name{projGridCP}
\alias{projGridCP}
\title{Project Subspaces for Contrasts, Penalty Parameters}
\usage{
projGridCP(target, center, scale, c_contrasts, contrasts, penalties,
  num_eigen)
}
\arguments{
\item{target}{The target data set}

\item{center}{Whether the data sets' columns should be centered}

\item{scale}{Whetehr the data sets' columns should be scaled to have variance}

\item{c_contrasts}{List of contrastve covariance matrices}

\item{contrasts}{Vector of contrast parameters}

\item{penalties}{Vector of penalty parameters}

\item{num_eigen}{The number of contrastive principal components to compute.}
}
\value{
A named list containing the loadings matrices, the projections of
  the target data into the spaces that they define and the grid of
  parameters.
}
\description{
Project the target data into the contrastive subspaces defined
  by the unique combinations of the contrast parameters and the penalty
  terms. All loading matrices that are kernels are dropped.
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
