% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scepca-class.R
\docType{class}
\name{scePCA-class}
\alias{scePCA-class}
\alias{.scePCA}
\title{Constructor for class scePCA}
\value{
class \code{scePCA} object, sub-classed from SingleCellExperiment.
}
\description{
Constructor for class scePCA
}
\examples{
library(SingleCellExperiment)
library(scPCA)

example_scePCA_class <- function(sce) {
  call <- match.call(expand.dots = TRUE)
  scepca <- .scePCA(
    SingleCellExperiment(
      assays = assay(sce),
      rowData = rowData(sce),
      colData = colData(sce)
    ),
    call = call,
    tmleOut = as.data.frame(matrix(NA, 10, 10)),
    topTable = as.data.frame(matrix(NA, 10, 10))
  )
  return(scepca)
}

example_class <- example_scepca_class(sce = EXAMPLEDATA)
}
