% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specClustSelection.R
\name{specClustSelection}
\alias{specClustSelection}
\title{Use Spectral Clustering to Select Contrastive Projections}
\usage{
specClustSelection(c_proj, num_medoids)
}
\arguments{
\item{c_proj}{The contrastive projection list produced by
\code{\link{projGridCP}}.}

\item{num_medoids}{The number of medoids to identify using spectral
clustering.}
}
\value{
A named list containing the matrix of parameters, the loadings matrix
  and the contrastive projection of the target associated with each medoid.
}
\description{
Use Spectral Clustering to Select Contrastive Projections
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
