% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{toy_df}
\alias{toy_df}
\title{Simulated Target Data for cPCA and scPCA}
\format{A simple \code{data.frame}.}
\usage{
data(toy_df)
}
\description{
The toy data consisting of 400 observations and 31 variables was simulated as
follows:
\itemize{
  \item Each of the first 10 variables was drawn from $N(0, 10)$
  \item For group 1 and 2, variables 11 through 20 were drawn from $N(0, 1)$
  \item For group 3 and 4, variables 11 through 20 were drawn from $N(3, 1)$
  \item For group 1 and 3, variables 21 though 30 were drawn from $N(-3, 1)$
  \item For group 2 and 4, variables 21 though 30 were drawn from $N(0, 1)$
  \item The last column provides each observations group number
}
}
\examples{
data(toy_df)

}
\keyword{datasets}
