% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastiveCov.R
\name{bpContrastiveCov}
\alias{bpContrastiveCov}
\title{Parallelized Contrastive Covariance Matrices}
\usage{
bpContrastiveCov(target, background, contrasts, center, scale)
}
\arguments{
\item{target}{The target data set.}

\item{background}{The background data set.}

\item{contrasts}{The vector of contrastive parameters.}

\item{center}{A \code{logical} indicating whether the data sets' columns
should be centered so as to have mean zero.}

\item{scale}{A \code{logical} indicating whether the data sets' columns
should be re-scaled to have unit variance.}
}
\value{
A list of contrastive covariance matrices. Each element has an
  associated contrastive parameter in the \code{contrasts} vector.
}
\description{
Compute the list of contrastive covariance matrices in parallel
  using \code{\link[BiocParallel]{bplapply}}.
}
