% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArgs.R
\name{checkArgs}
\alias{checkArgs}
\title{Check Arguments in scPCA Function}
\usage{
checkArgs(target, background, center, scale, num_eigen, contrasts,
  penalties)
}
\arguments{
\item{target}{Target data}

\item{background}{Background data}

\item{center}{Indicates whether target and background should be centered.}

\item{scale}{Indicates whether target and background should be scaled.}

\item{num_eigen}{The number of eigen vectors to compute.}

\item{contrasts}{The vector of contrasts parameters}

\item{penalties}{The vector of penalty terms}
}
\value{
Whether all argument conditions are satisfied
}
\description{
Checks whether or not the all arguments in the scPCA
functions are input properly.
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
