% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArgs.R
\name{checkArgs}
\alias{checkArgs}
\title{Check Arguments passed to the scPCA Function}
\usage{
checkArgs(
  target,
  background,
  center,
  scale,
  n_eigen,
  contrasts,
  penalties,
  clust_method,
  linkage_method
)
}
\arguments{
\item{target}{The target (experimental) data set, in a standard format such
as a \code{data.frame} or \code{matrix}.}

\item{background}{The background data set, in a standard format such as a
\code{data.frame} or \code{matrix}.}

\item{center}{A \code{logical} indicating whether the target and background
data sets should be centered to mean zero.}

\item{scale}{A \code{logical} indicating whether the target and background
data sets should be scaled to unit variance.}

\item{n_eigen}{A \code{numeric} indicating the number of eigenvectors to be
computed.}

\item{contrasts}{A \code{numeric} vector of the contrastive parameters.}

\item{penalties}{A \code{numeric} vector of the penalty terms.}

\item{clust_method}{A \code{character} specifying the clustering method to
use for choosing the optimal constrastive parameter. Currently, this is
limited to either k-means, partitioning around medoids (PAM), and
hierarchical clustering. The default is k-means clustering.}

\item{linkage_method}{A \code{character} specifying the agglomerative linkage
method to be used if \code{clust_method = "hclust"}. The options are
\code{ward.D2}, \code{single}, \code{complete},
\code{average}, \code{mcquitty}, \code{median}, and \code{centroid}. The
default is \code{complete}.}
}
\value{
Whether all argument conditions are satisfied
}
\description{
Checks whether or not the all arguments in the \code{scPCA}
  functions are input properly.
}
\keyword{internal}
