% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArgs.R
\name{checkArgs}
\alias{checkArgs}
\title{Check Arguments in cPCA and scPCA Functions}
\usage{
checkArgs(func, target, background, center, scale, num_eigen, contrasts,
  penalties, num_contrasts, num_medoids)
}
\arguments{
\item{target}{Target data}

\item{background}{Background data}

\item{center}{Indicates whether target and background should be centered.}

\item{scale}{Indicates whether target and background should be scaled.}

\item{num_eigen}{The number of eigen vectors to compute.}

\item{contrasts}{The vector of contrasts parameters}

\item{penalties}{The vector of penalty terms}

\item{num_medoids}{The number of medoids to find during spectral clustering.}

\item{which_func}{The function whose arguments are being checked. Either}
}
\value{
Whether all argument conditions are satisfied
}
\description{
Checks whether or not the all arguments in cPCA and scPCA
functions are input properly.
}
