% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceMatrix.R
\name{coerceMatrix}
\alias{coerceMatrix}
\title{Coerce Matrix Objects to Base Matrix Objects}
\usage{
coerceMatrix(data)
}
\arguments{
\item{data}{The data to be coerced to a matrix object.}
}
\value{
A coerced matrix object.
}
\description{
Coerces an object from classes in the \code{Matrix} package to
 the base matrix class.
}
\keyword{internal}
