% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMat.R
\name{covMat}
\alias{covMat}
\title{Compute Sample Covariance Matrix}
\usage{
covMat(data, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{The data with which to compute the sample covariance matrix.}

\item{center}{Indicates if the data should be centered to have mean 0.
Defaults to \code{TRUE}.}

\item{scale}{Indicates if the data should be scaled to have variance 1.
Defaults to \code{TRUE}.}
}
\value{
the covariance matrix of the data.
}
\description{
\code{covMat} computes the sample covariance matrix of a data
  set. If a variable in the dataset has zero variance, then its corresponding
  row and column in the covariance matrix are zero vectors.
}
\examples{
covMat(toy_df[, 2:31])
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
