% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPCA.R
\name{cPCA}
\alias{cPCA}
\title{Constrastive Principal Component Analysis}
\usage{
cPCA(target, background, num_eigen, contrast)
}
\arguments{
\item{target}{The target data. Either a dataframe or a matrix with
observations as rows and features as columns.}

\item{background}{The background data. Either a dataframe or a matrix with
observations as rows and features as columns. The number of features
must match the number of features in the target data.}

\item{num_eigen}{The number of contrastive principal components to compute.}

\item{contrast}{The value of the contrastive parameter.}
}
\value{
A list containing the lower-dimensional representation of the target
  data as a data frame and the column matrix of contrastive principal
  components.
}
\description{
Given target and background dataframes or matrices, \code{cPCA}
  will perform the contrastive principal component analysis of the target
  data for a given number of eigenvectors and a predetermined value
  for the contrast parameter. For more information on this method, see
  \href{https://www.nature.com/articles/s41467-018-04608-8#ref-CR29}
  {Abid et al.}.
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
