% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrid.R
\name{fitGrid}
\alias{fitGrid}
\title{Identify the Optimal Contrastive and Penalty Parameters}
\usage{
fitGrid(target, target_valid = NULL, center, scale, c_contrasts,
  contrasts, penalties, n_eigen, clust_method = c("kmeans", "pam"),
  n_centers, max_iter = 10)
}
\arguments{
\item{target}{The target (experimental) data set, in a standard format such
as a \code{data.frame} or \code{matrix}.}

\item{target_valid}{A holdout set of the target (experimental) data set, in a
standard format such as a \code{data.frame} or \code{matrix}. \code{NULL} by
default but used by \code{\link{cvSelectParams}} for cross-validated
selection of the contrastive and penalization parameters.}

\item{center}{A \code{logical} indicating whether the target and background
data sets should be centered to mean zero.}

\item{scale}{A \code{logical} indicating whether the target and background
data sets should be scaled to unit variance.}

\item{c_contrasts}{A \code{list} of contrastive covariances.}

\item{contrasts}{A \code{numeric} vector of the contrastive parameters used
to compute the contrastive covariances.}

\item{penalties}{A \code{numeric} vector of the penalty terms.}

\item{n_eigen}{A \code{numeric} indicating the number of eigenvectors to be
computed.}

\item{clust_method}{A \code{character} specifying the clustering method to
use for choosing the optimal constrastive parameter. Currently, this is
limited to either k-means or partitioning around medoids (PAM). The default
is k-means clustering.}

\item{n_centers}{A \code{numeric} giving the number of centers to use in the
clustering algorithm.}

\item{max_iter}{A \code{numeric} giving the maximum number of iterations to
be used in k-means clustering, defaulting to 10.}
}
\value{
A list similar to that output by \code{\link[stats]{prcomp}}:
  \itemize{
    \item rotation - the matrix of variable loadings
    \item x - the rotated data, centred and scaled, if requested, data
    multiplied by the rotation matrix
    \item contrast - the optimal contrastive parameter
    \item penalty - the optimal L1 penalty term
  }
}
\description{
This function is used to automatically select the optimal
  contrastive parameter and L1 penalty term for scPCA based on a clustering
  algorithm and average silhouette width.
}
\keyword{internal}
