% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrid.R
\name{fitGrid}
\alias{fitGrid}
\title{Identify the Optimal Contrastive and Penalty Parameters}
\usage{
fitGrid(target, center, scale, c_contrasts, contrasts, penalties, n_eigen,
  clust_method = c("kmeans", "pam"), n_centers, max_iters = 10)
}
\arguments{
\item{target}{The target data set.}

\item{center}{A \code{logical} indicating whether the data sets' columns
should be centered so as to have mean zero.}

\item{scale}{A \code{logical} indicating whether the data sets' columns
should be re-scaled to have unit variance.}

\item{c_contrasts}{List of of contrastive covariances.}

\item{contrasts}{Vector of contrastive parameter values used to compute the
contrastive covariances,}

\item{penalties}{Vector of penalty parameters.}

\item{n_eigen}{The number of contrastive principal components to compute.}

\item{clust_method}{A \code{character} specifying the clustering method to
use for choosing the optimal constrastive parameter. Currently, this is
limited to either k-means or partitioning around medoids (PAM).}

\item{n_centers}{The number of n_centers to use in the clustering algorithm.}

\item{max_iters}{The maximum number of iterations to use in k-means
clustering. Defaults to 10.}
}
\value{
A list similar to that output by \code{\link[stats]{prcomp}}:
  \itemize{
    \item rotation - the matrix of variable loadings
    \item x - the rotated data, centred and scaled, if requested, data
    multiplied by the rotation matrix
    \item contrast - the optimal contrastive parameter
    \item penalty - the optimal L1 penalty term
  }
}
\description{
This function is used to automatically select the optimal
  contrastive parameter and L1 penalty term for scPCA based on a clustering
  algorithm and average silhouette width.
}
\author{
Philippe Boileau, \email{philippe_Boileau@berkeley.edu}
}
