% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_scale.R
\name{safeColScale}
\alias{safeColScale}
\title{Safe Centering and Scaling of Columns}
\usage{
safeColScale(
  X,
  center = TRUE,
  scale = TRUE,
  tol = .Machine$double.eps,
  eps = 0.01
)
}
\arguments{
\item{X}{An input \code{matrix} to be centered and/or scaled.}

\item{center}{A \code{logical} indicating whether to re-center the columns
of the input matrix \code{X}.}

\item{scale}{A \code{logical} indicating whether to re-scale the columns of
the input matrix \code{X}.}

\item{tolVar}{A tolerance level for the lowest column variance (or standard
deviation) value to be tolerated when scaling is desired. The default is
set to \code{double.eps} of machine precision \code{\link[base]{.Machine}}.}

\item{epsVar}{The desired lower bound of the estimated variance for a given
column. When the lowest estimate falls below \code{tolVar}, it is truncated
to the value specified in this argument. The default is 0.01.}
}
\description{
This is a safer utility for centering and scaling an input matrix \code{X}.
It is intended to avoid some of the drawbacks of \code{\link[base]{scale}}
while simultaneously being faster through relying on \pkg{matrixStats} for
key internal computations.
}
\keyword{internal}
