% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPCA.R
\name{scPCA}
\alias{scPCA}
\title{Sparse Constrastive Principal Component Analysis}
\usage{
scPCA(target, background, center = TRUE, scale = FALSE, n_eigen = 2,
  contrasts = exp(seq(log(0.1), log(1000), length.out = 40)),
  penalties = seq(0.05, 1, length.out = 20), clust_method = "kmeans",
  n_centers, max_iters = 10, parallel = FALSE)
}
\arguments{
\item{target}{The target data. Either a numeric dataframe or a matrix with
observations as rows and features as columns.}

\item{background}{The background data. Either a numeric dataframe or a matrix
with observations as rows and features as columns. The number of features
must match the number of features in the target data.}

\item{center}{Whether the target and background data should have their
columns' centered. Defaults to \code{TRUE}.}

\item{scale}{Whether the target and background data should have their
columns' scaled. Defaults to \code{FALSE}.}

\item{n_eigen}{The number of sparse contrastive components to compute.
Default is 2.}

\item{contrasts}{The numeric vector of the contrastive parameters. Each
element must be a unique non-negative real number. Defaults to 40
logarithmically spaced values between 0.1 and 1000.}

\item{penalties}{The numeric vector of penatly terms for the L1 pernalty on
the loadings. Defaults to 20 equidistant values between  0.05 and 1.}

\item{clust_method}{A \code{character} specifying the clustering method to
use for choosing the optimal constrastive parameter. Currently, this is
limited to either k-means or partitioning around medoids (PAM). The default
is k-means.}

\item{n_centers}{The number of centers to use in the clustering algorithm.}

\item{max_iters}{The maximum number of iterations to use in k-means.
clustering. Defaults to 10.}

\item{parallel}{Boolean indicating whether parallel processing is used.
Defaults to \code{FALSE}.}
}
\value{
A list containing the following components:
  \itemize{
    \item rotation - the matrix of variable loadings
    \item x - the rotated data, centred and scaled if requested, multiplied
    by the rotation matrix
    \item contrast - the optimal contrastive parameter
    \item penalty - the optimal L1 penalty term
    \item center - whether the target dataset was centered
    \item scale - whether the target dataset was scaled
  }
}
\description{
Given target and background dataframes or matrices, \code{scPCA}
  will perform the sparse contrastive principal component analysis (scPCA) of
  the target data for a given number of eigenvectors, a vector of real valued
  contrast parameters and a vector of penalty terms. For more information on
  the contrastive PCA method, which this method is an extension of, consult
  \insertRef{abid2017contrastive}{scPCA}. Sparce PCA is performed using
  the method described in \insertRef{Zou2006}{scPCA}.
}
\examples{
# perform cPCA on the simulated data set
scPCA(
  target = toy_df[, 1:30],
  background = background_df,
  contrasts = exp(seq(log(0.1), log(100), length.out = 10)),
  penalties = 0,
  n_centers = 4
)

# perform scPCA on the simulated data set
scPCA(
  target = toy_df[, 1:30],
  background = background_df,
  contrasts = exp(seq(log(0.1), log(100), length.out = 10)),
  penalties = seq(0.1, 1, length.out = 9),
  n_centers = 4
)

# perform same operations in parallel
scPCA(
  target = toy_df[, 1:30],
  background = background_df,
  contrasts = exp(seq(log(0.1), log(100), length.out = 10)),
  penalties = 0,
  n_centers = 4,
  parallel = TRUE
)

# perform scPCA on the simulated data set
scPCA(
  target = toy_df[, 1:30],
  background = background_df,
  contrasts = exp(seq(log(0.1), log(100), length.out = 10)),
  penalties = seq(0.1, 1, length.out = 9),
  n_centers = 4,
  parallel = TRUE
)
}
\author{
Philippe Boileau, \email{philippe_boileau@berkeley.edu}
}
