% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPCA.R
\name{selectParams}
\alias{selectParams}
\title{Selection of Contrastive and Penalization Parameters}
\usage{
selectParams(target, background, center, scale, n_eigen, contrasts,
  penalties, clust_method, n_centers, max_iter, n_medoids, parallel)
}
\arguments{
\item{target}{The target (experimental) data set, in a standard format such
as a \code{data.frame} or \code{matrix}.}

\item{background}{The background data set, in a standard format such as a
\code{data.frame} or \code{matrix}. Note that the number of features must
match the number of features in the target data.}

\item{center}{A \code{logical} indicating whether the target and background
data sets should be centered to mean zero.}

\item{scale}{A \code{logical} indicating whether the target and background
data sets should be scaled to unit variance.}

\item{n_eigen}{A \code{numeric} indicating the number of eigenvectors (or
sparse contrastive components) to be computed. The default is to compute two
such eigenvectors.}

\item{contrasts}{A \code{numeric} vector of the contrastive parameters. Each
element must be a unique non-negative real number. The default is to use 40
logarithmically spaced values between 0.1 and 1000.}

\item{penalties}{A \code{numeric} vector of the L1 penalty terms on the
loadings. The default is to use 20 equidistant values between 0.05 and 1.}

\item{clust_method}{A \code{character} specifying the clustering method to
use for choosing the optimal constrastive parameter. Currently, this is
limited to either k-means or partitioning around medoids (PAM). The default
is k-means clustering.}

\item{n_centers}{A \code{numeric} giving the number of centers to use in the
clustering algorithm. If set to 1, cPCA, as first proposed by Abid et al.,
is performed, regardless of what the \code{penalties} argument is set to.}

\item{max_iter}{A \code{numeric} giving the maximum number of iterations to
be used in k-means clustering, defaulting to 10.}

\item{n_medoids}{A \code{numeric} indicating the number of medoids to
consider if \code{n_centers} is set to 1. The default is 8 such medoids.}

\item{parallel}{A \code{logical} indicating whether to invoke parallel
processing via the \pkg{BiocParallel} infrastructure. The default is
\code{FALSE} for sequential evaluation.}
}
\value{
Output structure matching either that of \code{\link{fitCPCA}} or
 \code{\link{fitGrid}} (or their parallelized variants, namely either
 \code{\link{bpFitCPCA}} and \code{link{bpFitGrid}}, respectively).
}
\description{
A wrapper function for fitting various internal functions to
 select the optimal setting of the contrastive and penalization parameters.
 For internal use only.
}
\keyword{internal}
